+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 0.12.0
 * Add support to release specific versions (#108)
 * Feature: Add possibilitity to tag release as Github prerelease (#104)
 * Bugfix: Ensure before and after tasks run correctly (#105)